% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_comp.R
\name{CI_ELR}
\alias{CI_ELR}
\title{A function that uses *uniroot* to solve for the ($1-\alpha$) confidence interval of sample mean
by the method of empirical likelihood}
\usage{
CI_ELR(y, alpha, ub)
}
\arguments{
\item{y}{samples}

\item{alpha}{significance level}

\item{ub}{upper bound for solving for the root}
}
\value{
a ($1-\alpha$) confidence interval by the method of empirical likelihood
}
\description{
A function that uses *uniroot* to solve for the ($1-\alpha$) confidence interval of sample mean
by the method of empirical likelihood
}
\details{
For empirical likelihood method, the ($1-\alpha$) CI for $\mu$ is
$$\{\mu|\sum_{i=1}^{n}log(np_i(\mu))>-0.5\chi^2_{1-\alpha}(1)\}$$
Noticing that $\mu$ doesn't have explicit formula to calculate, 
I solved the equation numerically.
}
\examples{
\dontrun{
set.seed(7524)
n=c(500,5000,50000)
lik=matrix(0,ncol=2,nrow=3)
rownames(lik)=paste("n=",n);colnames(lik)=c("left","right")
up=c(1.23,1.07,1.07)
for (i in 1:3){
 x=rnorm(n[i])
 y=x^2 #turn var into mean
 lik[i,]=CI_ELR(y,0.05,up[i])}
lik
}
}
