% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_comp.R
\name{CI_NDM}
\alias{CI_NDM}
\title{A function that computes the the ($1-\alpha$) confidence interval of sample mean
by the method of nonparametric delta method}
\usage{
CI_NDM(y, alpha)
}
\arguments{
\item{y}{samples}

\item{alpha}{significance level}
}
\value{
a ($1-\alpha$) confidence interval by the method of nonparametric delta method
}
\description{
A function that computes the the ($1-\alpha$) confidence interval of sample mean
by the method of nonparametric delta method
}
\details{
For nonparametric delta method of mean value, $\hat{\tau}^2=\frac{1}{n}\sum_{i=1}^{n}(Y_i-\bar{Y})^2$
 $$[\bar{Y}-\frac{z_{1-\frac{\alpha}{2}}*\tau}{\sqrt{n}},\bar{Y}+-\frac{z_{1-\frac{\alpha}{2}}*\tau}{\sqrt{n}}]$$
}
\examples{
\dontrun{
set.seed(7524)
n=c(500,5000,50000)
int=matrix(0,ncol=2,nrow=3)
rownames(int)=paste("n=",n);colnames(int)=c("left","right")
for (i in 1:3){
 x=rnorm(n[i])
 y=x^2 #turn var into mean
 int[i,]=CI_NDM(y,0.05)}
int
}
}
