% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{count5_test}
\alias{count5_test}
\title{A permutation test for equal variance based on the maximum number of extreme points}
\usage{
count5_test(y, n1)
}
\arguments{
\item{y}{the samples}

\item{n1}{the length of the first sample}
}
\value{
1 for reject null hypothesis (equal variance), 0 for accept
}
\description{
A permutation test for equal variance based on the maximum number of extreme points
}
\examples{
\dontrun{
compute the typr-I error for the permutation test
n1=20;n2=30;m=500
a2 = mean(replicate(m, expr={ 
x1 = rnorm(n1);x2 = rnorm(n2)
y=c(x1,x2)
count5_test(y,n1)
}))
}
}
