% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{count5num}
\alias{count5num}
\title{Compute the maximum number of extreme points}
\usage{
count5num(z, ix, n1)
}
\arguments{
\item{z}{the samples}

\item{ix}{the index after permutation}

\item{n1}{the length of the first sample}
}
\value{
the number of extreme points
}
\description{
Compute the maximum number of extreme points
}
