% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{count5test}
\alias{count5test}
\title{The original "Count Five” test for equal variance}
\usage{
count5test(x, y)
}
\arguments{
\item{x}{sample 1}

\item{y}{sample 2}
}
\value{
1 for reject null hypothesis (equal variance), 0 for accept
}
\description{
The original "Count Five” test for equal variance
}
\examples{
\dontrun{
compute the type-I error for the original test
n1=20;n2=30;m=500
a1 = mean(replicate(m, expr={ 
 x = rnorm(n1);y <- rnorm(n2)
 x = x - mean(x) #centered by sample mean
 y = y - mean(y)
 count5test(x, y)
}))
}
}
