% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{ff}
\alias{ff}
\title{Compute frac(2*Gamma(frac(k+1,2)),sqrt(pi*k)*Gamma(frac(k,2))*integral((1+frac(u^2,k))^{-(k+1)/2}*du,0,c[k])}
\usage{
ff(k, a)
}
\arguments{
\item{k}{integer}

\item{a}{real number}
}
\value{
the numerical integration
}
\description{
c[k]=sqrt(frac(a^2*k,k+1-a^2))
}
