% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{gen_Mixnorm}
\alias{gen_Mixnorm}
\title{Histogram of mixture of normal distribution}
\usage{
gen_Mixnorm(p1)
}
\arguments{
\item{p1}{The proportion of mixture}
}
\value{
A histogram of the sample with density superimposed
}
\description{
Generate a random sample of size 1000 from a normal location mixture. 
X[1]%~% N(0,1),X[2]%~% N(3,1),Z=p[1]*X[1]+(1-p[1])*X[2]
Graph the histogram of the sample with density superimposed
}
\examples{
\dontrun{
 gen_Mixnorm(0.25)
}
}
