% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jh_extend.R
\name{jh.tie}
\alias{jh.tie}
\title{Extended jonckheere function that can handle data with ties.}
\usage{
jh.tie(x, gr)
}
\arguments{
\item{x}{the multi-group data (vector)}

\item{gr}{the group index for the data (numeric/factor)}
}
\value{
a list that contains the value of J statistics and p.value
}
\description{
Extended jonckheere function that can handle data with ties.
}
\details{
For Jonkheere-Terpstra test with ties
$$W_{ij}=\#(X_{ik}<X_{jl},k=1,2,...,n_i,l=1,2,...,n_j)+\frac{1}{2}\#(X_{ik}=X_{jl},k=1,2,...,n_i,l=1,2,...,n_j)$$
$$J=\sum_{i<J}W_{ij}$$
$$\frac{J-E_{H_0}(J)}{\sqrt{var_{H_0}(J)}}$$
$E_{H_0}(J)=\frac{N^2-\sum_{i=1}^{k}n_i^2}{4}$
}
\examples{
\dontrun{
 data(m)
 attach(m)
 jh.tie(medical,gr)
}
}
