% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jh_extend.R
\name{jh_extend}
\alias{jh_extend}
\title{Jonckheere function that can handle data both with ties and without ties.}
\usage{
jh_extend(x, gr)
}
\arguments{
\item{x}{the multi-group data (vector)}

\item{gr}{the group index for the data (numeric/factor)}
}
\value{
a list that contains the value of J statistics and p.value
}
\description{
Jonckheere function that can handle data both with ties and without ties.
}
\examples{
\dontrun{
 data(m)
 attach(m)
 jh_extend(medical,gr)
}
}
