% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_comp.R
\name{log_npi}
\alias{log_npi}
\title{A function used to calculate the log-likelihood of sample mean}
\usage{
log_npi(x, mu)
}
\arguments{
\item{x}{samples}

\item{mu}{theoretical mean}
}
\value{
the value of $\sum_{i=1}^{n}log(np_i(\mu))$
}
\description{
A function used to calculate the log-likelihood of sample mean
}
\details{
$$When\ n\ is\ large,\ np_i\approx \frac{1}{1+\frac{(\bar{y}-\mu)(y_i-\mu)}{S(\mu)}}$$
where$S(\mu)=\frac{1}{n}\sum_{i=1}^{n}(y_i-\mu)^2$
}
