% Generated by  hand
\name{m}
\docType{data}
\alias{m}
\title{A data example for Jonckheere-Terpstra Test}
\description{
This data set gives an example to be used in jh.extend function.
}
\usage{m}
\format{A dataframe containing data and the group index}
\source{Xing Wang,Nonparametric Statistics}
\keyword{datasets}
