% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{MC.Phi}
\alias{MC.Phi}
\title{Monte Carlo integration with antithetic variables}
\usage{
MC.Phi(R = 10000, anti = TRUE)
}
\arguments{
\item{R}{number of generated samples to estimate the integration}

\item{anti}{use antithetic variables or not}
}
\value{
The numerical estimation of $\int_{0}^{1}\frac{e^{-x}}{1+x^2}dx$
}
\description{
This function is defined to estimate $\int_{0}^{1}\frac{e^{-x}}{1+x^2}dx$
and test the effect of antithetic variables in variance reduction.
}
\examples{
\dontrun{
set.seed(2222)
m=1000;
MC.de=MC=numeric(m)
for (i in 1:m){
 MC.de[i]=MC.Phi(R=1000)
 MC[i]=MC.Phi(R=1000,anti=FALSE)
 }
print("Approximate reduction in variance: ");(var(MC)-var(MC.de))/var(MC)
}
}
