% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{mcsapply}
\alias{mcsapply}
\title{A multicore version of sapply() (For Windows only)}
\usage{
mcsapply(cluster, X, FUN, ...)
}
\arguments{
\item{cluster}{}

\item{X}{object of the function}

\item{FUN}{function}

\item{...}{}
}
\value{

}
\description{
A multicore version of sapply() (For Windows only)
}
\examples{
\dontrun{
cores = detectCores()
cluster = makePSOCKcluster(cores)
boot_i = function(i){
r_adj = function(x,id) {
  x = x[id,]     #bootstrap sample
  res = lm(mpg~wt+disp,data=x)
  summary(res)$adj.r.squared
}
as.numeric(boot::boot(mtcars,r_adj,1)$t)
}
system.time(sapply(1:500,boot_i))
system.time(mcsapply(cluster,1:500,boot_i))
}
}
