% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{pwr_beta}
\alias{pwr_beta}
\title{The graph of the power of the skewness test of symmetric Beta$(\alpha, \alpha)$ distribution}
\usage{
pwr_beta(a)
}
\arguments{
\item{a}{the parameter for symmetric Beta distribution}
}
\value{
a graph of the power of the skewness test
}
\description{
The graph of the power of the skewness test of symmetric Beta$(\alpha, \alpha)$ distribution
}
\examples{
\dontrun{
 a1 = c(seq(0,1,0.1))
 pwr_beta(a1)
 a2 = c(seq(1,20,2),seq(20,100,5))
 pwr_beta(a2)
}
}
