% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{pwr_t}
\alias{pwr_t}
\title{The graph of the power of the skewness test of t distribution}
\usage{
pwr_t(v)
}
\arguments{
\item{v}{the parameter of t distribution}
}
\value{
a graph of the power of the skewness test
}
\description{
The graph of the power of the skewness test of t distribution
}
\examples{
\dontrun{
v=seq(1,20);pwr_t(v)
}
}
