% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{random.M}
\alias{random.M}
\title{A random walk Metropolis sampler for generating the standard Laplace distribution}
\usage{
random.M(sigma, x0, N)
}
\arguments{
\item{sigma}{the standard deviation of the sample-generating normal distribution}

\item{x0}{Initial values for the random walk}

\item{N}{number of the generated samples}
}
\value{
alist of a Markov Chain sampled from standard Laplace distribution and the times of rejection
}
\description{
A random walk Metropolis sampler for generating the standard Laplace distribution
}
\examples{
\dontrun{
N=2000
sigma = c(.05, .5, 2, 16)
x0=20
R1=random.M(sigma[1],x0,N)
R2=random.M(sigma[2],x0,N)
R3=random.M(sigma[3],x0,N)
R4=random.M(sigma[4],x0,N)
}
}
