% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{randomMHC}
\alias{randomMHC}
\title{A Metropolis-Hastings sampler using Rcpp}
\usage{
randomMHC(sigma, x0, N)
}
\arguments{
\item{sigma}{the standard deviation of the sample-generating normal distribution}

\item{x0}{initial values}

\item{N}{length of Markov Chain}
}
\value{
A Markov China of generating samples
}
\description{
A Metropolis-Hastings sampler using Rcpp
}
\examples{
\dontrun{
R= randomMHC(2,10,2000)
}
}
