\name{SC19090-package}
\alias{SC19090-package}
\alias{SC19090}
\docType{package}
\title{
  Final Project for Statistical Computing course, 2019FALL, USTC
}
\description{
  This package includes two personalize function for Jonckheere-Terpstra test and Nonparametric confidence interval. It also includes all the homework answers and some basic R functions.
}
\details{
  All the functions are equipped with examples and theoretical details.
}
\author{
Xixu Hu, xixuhu1998@gmail.com.

Maintainer: Xixu Hu <xixuhu1998@gmail.com>
}
\references{
  Maria L. Rizzo. Statistical Computing with R
  Xing Wang. Nonparametric Statistics
  Hadley Wickham. Advanced R
}
\keyword{ Statistical Computing with R }

