% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{sd_sk}
\alias{sd_sk}
\title{Calculate the standard error of q sample quantile using exact variance formula of skewness}
\usage{
sd_sk(x, q, n)
}
\arguments{
\item{x}{the quantiles of skewness coefficient}

\item{q}{quantiles}

\item{n}{number of samples}
}
\value{
the standard error of q sample quantile of skewness coefficient
}
\description{
Calculate the standard error of q sample quantile using exact variance formula of skewness
}
\details{
The variance of q sample quantile of skewness is $Var(\hat{x_q})=\frac{q(1-q)}{nf^2(x)}$
where f is the density of the sampled distribution. 
Using the normal approximation for the density, the skewness is asymptotically normal with mean 0 and variance $\frac{6(n-2)}{(n+1)(n+3)}$
}
\examples{
\dontrun{
 m=1000;n=100
 stat = replicate(m, expr={
 x = rnorm(n)
 y = sk(x)
 })
 q = c(.025,.05,.95,.975)
 x = quantile(stat,q)
 print('The standard error of the estimates:');sd_sk(x,q,n)
}
}
