% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{sk}
\alias{sk}
\title{Computes the sample skewness coefficient}
\usage{
sk(x)
}
\arguments{
\item{x}{samples}
}
\value{
the sample skewness coefficient
}
\description{
Computes the sample skewness coefficient
}
\examples{
\dontrun{
 x=rnorm(100);skewness=sk(x)
}
}
