% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jh_extend.R
\name{jh.tie}
\alias{jh.tie}
\title{Extended jonckheere function that can handle data with ties.}
\usage{
jh.tie(x, gr)
}
\arguments{
\item{x}{the multi-group data (vector)}

\item{gr}{the group index for the data (numeric/factor)}
}
\value{
a list that contains the value of J statistics and p.value
}
\description{
Extended jonckheere function that can handle data with ties.
}
\details{
For Jonkheere-Terpstra test with ties
W[ij]=\#(X[ik]<X[jl],k=1,2,...,n[i],l=1,2,...,n[j])+frac(1,2)\#(X[ik]=X[jl],k=1,2,...,n[i],l=1,2,...,n[j])
J=sum(W[ij],i<j)
frac(J-E[H[0]](J),sqrt(var[H[0]](J)))
E[H[0]](J)=frac(N^2-sum(n[i]^2,i==1,k),4)
}
\examples{
\dontrun{
 data(m)
 attach(m)
 jh.tie(medical,gr)
}
}
