% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.R
\name{solve_eq}
\alias{solve_eq}
\title{Solve the equation of $\frac{2\Gamma(\frac{k}{2})}{\sqrt{\pi (k-1)} \Gamma(\frac{k-1}{2})}\int_0^{c_{k-1}}(1+\frac{u^2}{k-1})^{-k/2}du$
$=\frac{2\Gamma(\frac{k+1}{2})}{\sqrt{\pi k} \Gamma(\frac{k}{2})}\int_0^{c_{k}}(1+\frac{u^2}{k})^{-(k+1)/2}du$}
\usage{
solve_eq(k)
}
\arguments{
\item{k}{integer}
}
\value{
the root of the equation
}
\description{
Solve the equation of $\frac{2\Gamma(\frac{k}{2})}{\sqrt{\pi (k-1)} \Gamma(\frac{k-1}{2})}\int_0^{c_{k-1}}(1+\frac{u^2}{k-1})^{-k/2}du$
$=\frac{2\Gamma(\frac{k+1}{2})}{\sqrt{\pi k} \Gamma(\frac{k}{2})}\int_0^{c_{k}}(1+\frac{u^2}{k})^{-(k+1)/2}du$
}
\examples{
\dontrun{
Root=matrix(0,2,5)
k=c(4,25,100,500,1000)
for (i in 1:length(k)){
 Root[2,i]=round(solve_eq(k[i]),4)
}
Root[1,]=k;rownames(Root)=c('k','root')
Root
}
}
